requirements.txt
streamlit
sentence-transformers
scikit-learn
pdfplumber
python-docx
google-generativeai
torch